pulled out interface TensorValidator
