+++ keanu-project/src/test/java/io/improbable/keanu/tensor/dbl/TensorValidator.java
@@ -2,17 +2,6 @@ package io.improbable.keanu.tensor.dbl;
 
 import io.improbable.keanu.tensor.bool.BooleanTensor;
 
-public class TensorValidator {
-
-    private final Double value;
-
-    public TensorValidator(Double value) {
-
-        this.value = value;
-    }
-
-    public BooleanTensor check(DoubleTensor tensor) {
-        DoubleTensor testTensor = DoubleTensor.create(value, tensor.getShape());
-        return tensor.elementwiseEquals(testTensor);
-    }
+public interface TensorValidator {
+    BooleanTensor check(DoubleTensor tensor);
 }
